Implemented OutputSetItems option
This allows writing the counter values of individual
set items instead of the sum of all items in the set.

With the current Metrics.NET API only Meters and Counters
support this feature.

The option is disabled by default and can be enabled by
passing an options object to the WithPrometheusEndpointReport extension methods.